Move logging setup to separate file and execute before loading app
